Resolve branch-sdk server incompatibility
