+++ src/Client.js
@@ -18,7 +18,7 @@ import reducers from 'app/reducers';
 import reduxMiddleware from 'app/reduxMiddleware';
 import { sendTimings, onHandlerCompleteTimings } from 'lib/timing';
 import Session from 'app/models/Session';
-import * as smartBannerActions from 'app/actions/smartBanner';
+import * as smartBannerActions from 'app/actions/smartBannerClientOnly';
 
 // Bits to help in the gathering of client side timings to relay back
 // to the server
