+++ src/config.js
@@ -95,6 +95,8 @@ const config = () => ({
 
   // Note that this is a public key, so this can be shared.
   recaptchaSitekey: process.env.RECAPTCHA_SITEKEY || '6LeTnxkTAAAAAN9QEuDZRpn90WwKk_R1TRW_g-JC',
+
+  branchKey: process.env.BRANCH_KEY,
 });
 
 export default config();
